refactor(mga): extract UnplaceableLineupsNotice to its own file
The Task 7 hint was added as an inline MapPage sub-component. Extract it
to components/lineup/UnplaceableLineupsNotice.tsx (one component per
file) and add a dedicated render test. Pre-existing MapPage sub-components
and the pre-existing results-panel ternary are intentionally left as-is
(out of scope for this two-bug fix).
