+++ apps/mygamingassistant/frontend/src/pages/MapPage.tsx
@@ -31,6 +31,7 @@ import MapLineupPins, {
 } from "@/components/lineup/MapLineupPins";
 import PinModeToggle from "@/components/lineup/PinModeToggle";
 import LineupDetailPanel from "@/components/lineup/LineupDetailPanel";
+import UnplaceableLineupsNotice from "@/components/lineup/UnplaceableLineupsNotice";
 import KeyboardShortcutsHelp from "@/components/lineup/KeyboardShortcutsHelp";
 import MinimapUploadDialog from "@/components/game/MinimapUploadDialog";
 import RoundMode from "@/pages/RoundMode";
@@ -755,34 +756,6 @@ function StorageUnavailableBanner({ onClose }: { onClose: () => void }) {
       >
         ✕
       </button>
-    </div>
-  );
-}
-
-interface UnplaceableLineupsNoticeProps {
-  count: number;
-}
-
-function UnplaceableLineupsNotice({ count }: UnplaceableLineupsNoticeProps) {
-  const plural = count !== 1 ? "s" : "";
-  return (
-    <div
-      className="flex items-start gap-2.5 px-3 py-2.5 rounded-md border bg-amber-500/10 border-amber-500/30 text-sm"
-      role="status"
-    >
-      <AlertTriangle
-        className="w-4 h-4 text-amber-600 dark:text-amber-400 flex-shrink-0 mt-0.5"
-        aria-hidden
-      />
-      <div className="flex-1">
-        <p className="font-medium">
-          {count} lineup{plural} can't be shown on the map yet
-        </p>
-        <p className="text-xs text-muted-foreground">
-          These zones need calibration — open the lineups in Review or the
-          zone editor to set their map position.
-        </p>
-      </div>
     </div>
   );
 }
